+++ packages/manager/cypress/integration/linodes/smoke-linode-landing-table.spec.ts
@@ -11,6 +11,7 @@ import { accountSettingsFactory } from '@src/factories/accountSettings';
 import { routes } from 'cypress/support/ui/constants';
 import * as Factory from 'factory.ts';
 import { Linode } from '@linode/api-v4/lib/linodes/types';
+import { interceptOnce } from 'cypress/support/ui/common';
 
 const regions: string[] = [
   'us-east',
@@ -23,6 +24,9 @@ const mockLinodes = makeResourcePage(linodeFactory.buildList(5));
 mockLinodes.data.forEach((linode, index) => (linode.region = regions[index]));
 
 const appRoot = Cypress.env('REACT_APP_APP_ROOT');
+const linodeLabel = (number) => {
+  return mockLinodes.data[number - 1].label;
+};
 
 const deleteLinodeFromActionMenu = (linodeLabel) => {
   getClick(`[aria-label="Action menu for Linode ${linodeLabel}"]`);
@@ -38,6 +42,18 @@ describe('linode landing checks', () => {
     const mockAccountSettings = accountSettingsFactory.build({
       managed: false,
     });
+
+    interceptOnce('GET', '*/profile/preferences*', {
+      linodes_view_style: 'grid',
+      linodes_group_by_tag: true,
+      volumes_group_by_tag: false,
+      desktop_sidebar_open: false,
+      is_large_account: true,
+      sortKeys: {
+        'linodes-landing': { order: 'desc', orderBy: 'label' },
+        volume: { order: 'desc', orderBy: 'label' },
+      },
+    }).as('getProfilePreferences');
 
     cy.intercept('GET', '*/account/settings', (req) => {
       req.reply(mockAccountSettings);
@@ -128,6 +144,82 @@ describe('linode landing checks', () => {
     getClick('[aria-label="Sort by label"]');
     checkFirstRow(lastLinodeLabel);
     checkLastRow(firstLinodeLabel);
+  });
+
+  it('checks the create menu dropdown items', () => {
+    getClick('[data-qa-add-new-menu-button="true"]');
+    getVisible(
+      '[data-valuetext="LinodeHigh performance SSD Linux servers"][href="/linodes/create"]'
+    );
+    getVisible(
+      '[data-valuetext="VolumeAttach additional storage to your Linode"][href="/volumes/create"]'
+    );
+    getVisible(
+      '[data-valuetext="NodeBalancerEnsure your services are highly available"][href="/nodebalancers/create"]'
+    );
+    getVisible(
+      '[data-valuetext="FirewallControl network access to your Linodes"][href="/firewalls/create"]'
+    );
+    getVisible(
+      '[data-valuetext="DomainManage your DNS records"][href="/domains/create"]'
+    );
+    getVisible(
+      '[data-valuetext="KubernetesHighly available container workloads"][href="/kubernetes/create"]'
+    );
+    getVisible(
+      '[data-valuetext="BucketS3-compatible object storage"][href="/object-storage/buckets/create"]'
+    );
+    getVisible(
+      '[data-valuetext="MarketplaceDeploy applications with ease"][href="/linodes/create?type=One-Click"]'
+    );
+  });
+
+  it('checks the table and action menu buttons/labels', () => {
+    const label = linodeLabel(1);
+    const ip = mockLinodes.data[0].ipv4[0];
+    getVisible('[aria-label="Sort by label"]').within(() => {
+      fbtVisible('Label');
+    });
+    getVisible('[aria-label="Sort by _statusPriority"]').within(() => {
+      fbtVisible('Status');
+    });
+    getVisible('[aria-label="Sort by type"]').within(() => {
+      fbtVisible('Plan');
+    });
+    getVisible('[aria-label="Sort by ipv4[0]"]').within(() => {
+      fbtVisible('IP Address');
+    });
+    getVisible(
+      '[aria-label="Toggle display"][aria-describedby="displayViewDescription"][title="Summary view"]'
+    );
+    getVisible(
+      '[aria-label="Toggle group by tag"][aria-describedby="groupByDescription"][title="Group by tag"]'
+    );
+
+    getVisible(`tr[data-qa-linode="${label}"]`).within(() => {
+      cy.get(`[data-qa-ip-main]`)
+        .realHover()
+        .then(() => {
+          getVisible(`[aria-label="Copy ${ip} to clipboard"]`);
+        });
+      getVisible(`[aria-label="Action menu for Linode ${label}"]`);
+    });
+  });
+
+  it('checks the action menu items', () => {
+    const label = linodeLabel(1);
+    getVisible(`tr[data-qa-linode="${label}"]`).within(() => {
+      fbtVisible('Power Off');
+      fbtVisible('Reboot');
+      getClick(`[aria-label="Action menu for Linode ${label}"]`);
+    });
+    getVisible('[data-qa-action-menu-item="Launch LISH Console"]');
+    getVisible('[data-qa-action-menu-item="Clone"]');
+    getVisible('[data-qa-action-menu-item="Resize"]');
+    getVisible('[data-qa-action-menu-item="Rebuild"]');
+    getVisible('[data-qa-action-menu-item="Rescue"]');
+    getVisible('[data-qa-action-menu-item="Migrate"]');
+    getVisible('[data-qa-action-menu-item="Delete"]');
   });
 });