+++ packages/manager/src/featureFlags.ts
@@ -37,6 +37,8 @@ export interface Flags {
   imagesPricingCopy: ImagesPricingCopy;
   referralBannerText: ReferralBannerText;
   blockStorageAvailability: boolean;
+  imagesPriceInfo: boolean;
+  apiMaintenance: string[];
 }
 
 type PromotionalOfferFeature =